+++ lib/loxxy/datatype/all_datatypes.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'false'
-require_relative 'true'
+require_relative 'lx_string'
+require_relative 'nil'
 require_relative 'number'
+require_relative 'true'