+++ CHANGELOG.md
@@ -1,4 +1,16 @@
+## [0.0.3] - 2020-12-29
+- Scanner can recognize strings and nil tokens
+### Added
+- class `Datatype::Nil` for representing nil tokens in Lox
+- class `Datatype::LXString` for representing Lox strings
+
+### Changed
+- Method `BuilinDatatype::validated_value` now accepts the input value as is.
+- `Frontend::Scanner` updated code to recognize strings and nil
+- File `scanner_spec.rb` Added tests for strings, nil scanning
+
 ## [0.0.2] - 2020-12-28
+- Scanner can recognize keywords, boolean and numbers
 ### Added
 - `Datatype` module and classes